wpt_babel/content-security-policy/script-src-attr-elem/script-src-elem-allowed-attr-blocked.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="script-src-elem 'self' 'unsafe-inline';
    script-src-attr 'none';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

var t = async_test("Should fire a security policy violation for the attribute");
window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
  assert_equals(e.violatedDirective, 'script-src-attr');
  assert_equals(e.blockedURI, 'inline');
}));
var t1 = async_test("Should execute the inline script block");
  </script>
  <script>
   "use strict";

t1.done();
  </script>
  <img onload="t.unreached_func('should not have run this event handler')" src="../support/pass.png"/>
 </body>
</html>
